Current settings now logged when provisioning
Copies current settings file to vlad_aux/tmp/current_build_settings.yml when provisioning. This is to provide an easy reminder of how the current VM was provisioned (a settings file can be altered but then not used after a VM was provisioned and so may not be an accurate indication of how the VM was set up).

Did a tiny bit of tidying up in the base role while I was there.