Use an extra var for inventory directory.
This does away with the concept of inventory directory entirely. After all:

- We are operating locally.
- The Vagrantfile already knows what directory we are in. We use this value elsewhere
- So why not just pass it into Ansible?
- The trick, then, is that we need to shuttle that variable from the host scope to our
local 127.0.0.1 scope. We do that with delegate_to.

I have tested this, and it works great! I'm not sure if this is as clean of a fix as
you want, but it's a better stopgap than what we have now. Refs #324.